Extract OpenId4VpRequestFactory shared by both verifiers
